CSS, JS, config and templated updates
