+++ app/assets/stylesheets/application.scss
@@ -1,0 +1,23 @@
+@import "lib/govuk-frontend/govuk/all";
+
+$govuk-global-styles: true;
+
+// ----------------
+// Hide the back link when body does not have .js-enabled
+body:not(.js-enabled) {
+    .govuk-back-link {
+        display: none;
+        visibility: hidden;
+        width: 0;
+        height: 0;
+    }
+}
+// ----------------
+
+p {
+    @extend .govuk-body;
+}
+
+a {
+    @extend .govuk-link;
+}
