--- app/assets/stylesheets/colours/_palette.scss
@@ -1,14 +1,0 @@
-// From https://github.com/alphagov/govuk_frontend_toolkit/blob/03204449dcbf28c2d48e7f6dd217f6abe9b3a518/stylesheets/colours/_palette.scss
-
-$govuk-text-colour: #0b0c0c;
-$white: #fff;
-$blue: #005ea5;
-$grey-1: #6f777b;
-$grey-2: #bfc1c3;
-$border-colour: $grey-2;
-$light-grey: #f3f2f1;
-$govuk-link-colour: #1d70b8;
-$yellow: #ffdd00;
-$green: #00703c;
-$red: #d4351c;
-$input-border: #0b0c0c;
