--- app/assets/stylesheets/govuk-colours-override.scss
@@ -1,262 +1,0 @@
-@import 'colours/palette';
-
-// =========================
-//
-// This file overwrites css rules in assets-frontend 3.11.0 to bring them inline with the
-// new colours in govuk design system
-// https://design-system.service.gov.uk/styles/colour/
-//
-// updated 3 Oct 2019, 19:27 BST
-// Palette change
-// https://designnotes.blog.gov.uk/2019/07/29/weve-updated-the-gov-uk-colours-and-font/
-// =========================
-
-// #DEE0E2 to #f3f2f1
-.button--secondary,
-.button--secondary:visited,
-.button--secondary.disabled:hover,
-.button--secondary[disabled="disabled"]:hover,
-.button--secondary[disabled]:hover,
-.footer,
-.box--shaded,
-.side-nav__link:hover,
-.dataTable tbody .status--unconfirmed,
-.tag,
-.tabs-nav .tabs-nav__tab:focus,
-.tabs-nav .tabs-nav__tab:hover,
-.tabs-nav .tabs-nav__tab:active,
-.tabs-nav a.tabs-nav__tab:focus,
-.tabs-nav .tabs-nav__tab.link-style:focus,
-.tabs-nav .tabs-nav__tab.button--link-style:focus,
-.tabs-nav a.tabs-nav__tab:hover,
-.tabs-nav .tabs-nav__tab.link-style:hover,
-.tabs-nav .tabs-nav__tab.button--link-style:hover,
-.tabs-nav a.tabs-nav__tab:active,
-.tabs-nav .tabs-nav__tab.link-style:active,
-.tabs-nav .tabs-nav__tab.button--link-style:active,
-.annual-summaries--group .statement__summary--parent[aria-expanded=true],
-.payments .table-label.selected,
-.payments .table-label:hover {
-  background-color: $light-grey;
-}
-.button--secondary:active {
-  -webkit-box-shadow: 0 0 0 $light-grey;
-  box-shadow: 0 0 0 $light-grey
-}
-.dynamically-displayed {
-  border-left: 5px solid $light-grey;
-}
-.explore a,
-.explore--inside-government a,
-.explore--inline a,
-.explore .link-style,
-.explore--inside-government .link-style,
-.explore--inline .link-style,
-.explore .button--link-style,
-.explore--inside-government .button--link-style,
-.explore--inline .button--link-style,
-.explore--inline__links {
-  color: $light-grey;
-}
-.indent {
-  border-left-color: $light-grey;
-}
-.services-dashboard .service-option--links:first-child .service-option--item,
-.account-menu__link:hover,
-.is-smaller .account-menu__link--back {
-  border-bottom-color: $light-grey;
-}
-
-
-// #005ea5 to #1d70b8
-.heading--underlined,
-.section--branded h2,
-.table--light-header thead,
-.table__header--highlight,
-.business-tax .section--branded h2,
-.is-smaller .account-menu__link--home.account-menu__link--active:hover,
-.account-menu__link--active,
-.account-menu__link--active:visited {
-  border-bottom-color: $govuk-link-colour;
-}
-.heading-large--blue,
-.button.button--link,
-a.back-to-top-link:visited,
-.back-to-top-link.link-style:visited,
-.back-to-top-link.button--link-style:visited,
-.page-nav__link,
-.banner-links,
-.side-nav--child .side-nav__list--selected .side-nav__link,
-.dataTables_paginate a,
-.dataTables_paginate .link-style,
-.dataTables_paginate .button--link-style,
-.accordion__button:visited,
-.toggle__button,
-.tabs-nav .tabs-nav__tab:visited,
-.tabs-nav__tab,
-.payments .tab,
-.services-dashboard .service-option--links .service-option--link:link,
-.link,
-.link:active,
-details summary,
-a,
-.link-style,
-.button--link-style,
-.button--link-style,
-.subnav a:visited,
-.subnav .link-style:visited,
-.subnav .button--link-style:visited,
-.is-smaller .subnav__section a:hover,
-.is-smaller .subnav__section .link-style:hover,
-.is-smaller .subnav__section .button--link-style:hover {
-  color: $govuk-link-colour;
-}
-.brand-border--top,
-.section--blue-top,
-.footer,
-.service-info {
-  border-top-color: $govuk-link-colour;
-}
-.button--notification,
-.button--notification:visited,
-.button--notification.disabled:hover,
-.button--notification[disabled="disabled"]:hover,
-.button--notification[disabled]:hover,
-.alpha-banner .phase-tag,
-.beta-banner .phase-tag,
-.phase-tag,
-.side-nav__list--selected .side-nav__link,
-.side-nav__list--selected:hover .side-nav__link,
-.highlight-message--dark,
-.hero,
-.paye-statements .sidebar-links__item.selected,
-option:active,
-option:checked,
-select:focus::-ms-value,
-.phase-banner .phase-tag,
-.phase-banner-alpha .phase-tag,
-.phase-banner-beta .phase-tag,
-.phase-tag,
-.full-width-banner,
-.hmrc-banner .phase-tag {
-  background-color: $govuk-link-colour;
-}
-.button--notification:active {
-  -webkit-box-shadow: 0 0 0 $govuk-link-colour;
-  box-shadow: 0 0 0 $govuk-link-colour;
-}
-.attorneyBanner,
-.panel-indent--info,
-.delegation-banner {
-  border-left-color: $govuk-link-colour;
-}
-
-// #ffbf47 to #ffdd00
-.page-nav__link:focus,
-.js-enabled .add-focus,
-.modal-dialog__content:focus,
-.button:focus,
-details summary:focus,
-.error-summary:focus {
-  outline-color: $yellow;
-}
-.multiple-choice [type=radio]:focus+label::before {
-  -webkit-box-shadow: 0 0 0 4px $yellow;
-  box-shadow: 0 0 0 4px $yellow;
-}
-.multiple-choice [type=checkbox]:focus+label::before {
-  -webkit-box-shadow: 0 0 0 3px $yellow;
-  box-shadow: 0 0 0 3px $yellow;
-}
-.account-menu__link:focus,
-.is-smaller .account-menu__link--home:focus,
-.is-smaller .account-menu__link--menu:focus,
-.is-smaller .account-menu__link--home.account-menu__link--active:focus,
-.is-smaller .account-menu__link--menu.account-menu__link--active:focus {
-  border-bottom-color: $yellow;
-}
-
-
-// #00823b to #00703c
-button,
-.button,
-button:visited,
-.button:visited,
-button.disabled:hover,
-button[disabled="disabled"]:hover,
-button[disabled]:hover,
-.button.disabled:hover,
-.button[disabled="disabled"]:hover,
-.button[disabled]:hover,
-.button--small,
-.button--small:visited,
-.button--small.disabled:hover,
-.button--small[disabled="disabled"]:hover,
-.button--small[disabled]:hover,
-.toggle__button--active,
-.toggle__button--active:hover,
-.toggle__button--active:focus,
-.button[disabled="disabled"] {
-  background-color: $green;
-}
-button:active,
-.button:active,
-.button--small:active {
-  -webkit-box-shadow: 0 0 0 $green;
-  box-shadow: 0 0 0 $green;
-}
-.list--ticks li::before,
-.tick,
-.form-hint-list-item--valid {
-  color: $green;
-}
-
-// #d4351c to #d4351c
-.button--alert,
-.button--alert:visited,
-.button--alert.disabled:hover,
-.button--alert[disabled="disabled"]:hover,
-.button--alert[disabled]:hover,
-.marker-column__marker,
-.reminder.urgent a,
-.reminder.urgent .link-style,
-.reminder.urgent .button--link-style,
-.alert-block,
-.alert--important,
-.tabs-nav .tabs-nav__tab .count,
-.flag.flag--urgent,
-.highlight-block--red,
-.main-nav .count,
-.badge {
-  background-color: $red;
-}
-.button--alert:active {
-  -webkit-box-shadow: 0 0 0 $red;
-  box-shadow: 0 0 0 $red;
-}
-.email-form-errors .error-notification,
-.annual-summaries--group .alert--text,
-.sent-in-error,
-.business-tax .content__body .flag.flag--urgent,
-.error-message,
-.error-summary .error-summary-list a,
-.error-summary .error-summary-list .link-style,
-.error-summary .error-summary-list .button--link-style {
-  color: $red;
-}
-
-.form-control.field-error {
-  border: 4px solid $red;
-}
-
-.email-form-errors .error-notification,
-.attorneyBanner--client-request-rejected,
-.business-tax .content__body .flag.flag--urgent,
-.form-group-error {
-  border-left-color: $red;
-}
-.business-tax .alert--failure,
-input.form-control-error[type="text"], textarea.form-control-error,
-.error-summary {
-  border-color: $red;
-}
