--- app/assets/stylesheets/maintainprotectorsfrontend-app.scss
@@ -1,237 +1,0 @@
-@import 'mixins/spacing';
-@import 'mixins/clearfix';
-
-@import 'partials/_variables';
-@import 'partials/_base';
-@import 'partials/_forms';
-@import 'partials/_typography';
-@import 'partials/_utilities';
-@import 'partials/_shame';
-@import 'partials/heading';
-@import 'patterns/tag';
-@import 'conditionals';
-@import 'colours/palette';
-
-// Autocomplete country lookup
-@import 'location-autocomplete.min';
-@import 'patterns/autocomplete';
-
-// WCGA 2.1 https://design-system.service.gov.uk/styles/colour/
-@import 'govuk-colours-override';
-
-@import 'partials/_focus';
-@import 'patterns/add-to-a-list';
-@import 'patterns/symbol-input';
-
-@import 'partials/task-list';
-@import 'partials/_modal';
-@import 'patterns/check-your-answers';
-
-.push-text-top--3 {
-  padding-top: govuk-spacing(3);
-}
-
-.field-error {
-  border: 4px solid $red !important;
-}
-
-.govuk-warning-text {
-  font-family: "nta", Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 1.25;
-  color: #0b0c0c;
-  position: relative;
-  margin-bottom: 20px;
-  padding: 10px 0;
-}
-
-@media (min-width: 40.0625em) {
-  .govuk-warning-text {
-    font-size: 19px;
-    line-height: 1.31579;
-  }
-}
-
-@media (min-width: 40.0625em) {
-  .govuk-warning-text {
-    margin-bottom: 30px;
-  }
-}
-
-.govuk-warning-text__icon {
-  font-family: "nta", Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  font-weight: 700;
-  display: inline-block;
-  position: absolute;
-  top: 50%;
-  left: 0;
-  min-width: 32px;
-  min-height: 29px;
-  margin-top: -20px;
-  padding-top: 3px;
-  border: 3px solid #0b0c0c;
-  border-radius: 50%;
-  color: #fff;
-  background: #0b0c0c;
-  font-size: 1.6em;
-  line-height: 29px;
-  text-align: center;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
-
-.govuk-warning-text__assistive {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  padding: 0;
-  overflow: hidden;
-  clip: rect(0 0 0 0);
-  -webkit-clip-path: inset(50%);
-  clip-path: inset(50%);
-  border: 0;
-  white-space: nowrap;
-}
-
-.govuk-warning-text__text {
-  display: block;
-  margin-left: -15px;
-  padding-left: 65px;
-  font-weight: bold;
-}
-
-.govuk-caption-xl {
-  font-family: "nta", Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  font-weight: 400;
-  font-size: 18px;
-  line-height: 1.11111;
-  display: block;
-  margin-bottom: 5px;
-  color: #6f777b;
-}
-
-@media print {
-  .govuk-caption-xl {
-    font-family: sans-serif;
-  }
-}
-
-@media (min-width: 40.0625em) {
-  .govuk-caption-xl {
-    font-size: 27px;
-    line-height: 1.11111;
-  }
-}
-
-@media print {
-  .govuk-caption-xl {
-    font-size: 18pt;
-    line-height: 1.15;
-  }
-}
-
-.govuk-fieldset__heading {
-  margin: 0;
-}
-
-.heading-no-top-margin {
-  margin-top: 0;
-}
-
-.label-as-heading, .legend-as-heading {
-  margin-top: 0;
-  margin-bottom: 0;
-}
-
-.heading-large, .heading-medium {
-  margin-top: 0;
-  word-break: break-word;
-}
-
-.heading-with-top-margin {
-  margin-top: .625em;
-}
-
-.heading-confirmation {
-  background-color: #00703c;
-  color: #fff;
-  padding: .6em;
-  text-align: center;
-  margin-top: 0;
-
-  .reference-text {
-    color: #fff;
-    margin-top: 1em;
-    display: block;
-    font-weight: 100;
-  }
-
-  .reference-number {
-    font-weight: 800;
-    font-size: 1em;
-    margin-top: 0;
-  }
-}
-
-.govuk-main-wrapper--l {
-  padding-top: 50px;
-
-  @include media(mobile) {
-    padding-top: 0;
-  }
-}
-
-.agentClientRef, .declarationReferenceNumber {
-  font-weight: 700;
-}
-
-.form-control.agent-internal-reference, .form-control.declaration-email {
-  width: 100%;
-}
-
-.error-message {
-  margin-bottom: 15px;
-}
-
-@media print {
-  .print-hidden, .header-wrapper, .service-info, .report-error {
-    display: none !important;
-  }
-  .content__body {
-    width: 100%;
-  }
-  .heading-xlarge {
-    font-size: 27px;
-  }
-  .heading-large {
-    font-size: 24px;
-  }
-  .heading-medium {
-    font-size: 19px;
-  }
-  ul li, p, .govuk-warning-text {
-    font-size: 16px;
-  }
-}
-
-.multiple-choice.first {
-  clear: both;
-}
-
-.govuk-details {
-  margin-bottom: 30px;
-}
-
-.clearfix {
-  @include govuk-clearfix;
-}
