--- app/assets/stylesheets/mixins/_clearfix.scss
@@ -1,15 +1,0 @@
-////
-/// @group helpers
-////
-
-/// Clear floated content within a container using a pseudo element
-///
-/// @access public
-
-@mixin govuk-clearfix {
-  &:after {
-    content: "";
-    display: block;
-    clear: both;
-  }
-}
