--- app/assets/stylesheets/partials/_base.scss
@@ -1,37 +1,0 @@
-/*
-====================================
-Fixes native-enforcing of date field for date-of-birth fields
-as service does not implement native date inputs
-====================================
-*/
-.touchevents .date-of-birth {
-  display:block;
-}
-
-
-
-/*
-====================================
-Work-around for issue on iOS VoiceOver with AFE version of visuallyhidden
-https://github.com/hmrc/assets-frontend/issues/727
-====================================
-*/
-.visually-hidden {
-position: absolute;
-overflow: hidden;
-clip: rect(0 0 0 0);
-height: 1px;
-width: 1px;
-margin: -1px;
-padding: 0;
-border: 0;
-}
-
-/*
-====================================
-Welsh/English language selector
-====================================
-*/
-.translate {
-  margin: 0;
-}
