--- app/assets/stylesheets/partials/_focus.scss
@@ -1,120 +1,0 @@
-// =========================
-// updated 29 Nov 2019 16:07 BST
-// focus states
-// changes made to bring in line with
-// https://designnotes.blog.gov.uk/2019/07/29/weve-made-the-gov-uk-design-system-more-accessible/
-// =========================
-
-  a:focus,
-  .error-summary .error-summary-list a:focus,
-  details summary:focus {
-    outline: 3px solid transparent;
-    color: #0b0c0c;
-    background-color: #fd0;
-    -webkit-box-shadow: 0 -2px #fd0, 0 4px #0b0c0c;
-    box-shadow: 0 -2px #fd0, 0 4px #0b0c0c;
-    text-decoration: none;
-  }
-  // error summary spacing
-  .error-summary-list li {
-    margin-bottom: 5px;
-  }
-  details summary:focus .summary {
-    text-decoration: none;
-  }
-  input:focus,
-  textarea:focus,
-  select:focus,
-  #global-header input:focus {
-    outline-color: #fd0;
-  }
-  .button {
-    border: 2px solid transparent;
-  }
-  .button:focus:not(:active):not(:hover) {
-    border-color: #fd0;
-    color: #0b0c0c;
-    background-color: #fd0;
-    -webkit-box-shadow: 0 2px 0 #0b0c0c;
-    box-shadow: 0 2px 0 #0b0c0c;
-  }
-  .button:focus {
-    border-color: #fd0;
-    outline: 3px solid transparent;
-    -webkit-box-shadow: inset 0 0 0 1px #fd0;
-    box-shadow: inset 0 0 0 1px #fd0;
-  }
-  .autocomplete__input:focus,
-  .form-control:focus {
-    outline: 3px solid #fd0;
-    outline-offset: 0;
-    -webkit-box-shadow: inset 0 0 0 2px;
-    box-shadow: inset 0 0 0 2px;
-  }
-  .multiple-choice [type=checkbox]:focus+label::before {
-    border-width: 3px;
-  }
-
-// header and footer (incl. account header)
-  .report-error .report-error__toggle:focus,
-  #footer a:focus,
-  #global-header .header-proposition #proposition-links li a:focus,
-  #global-header .header-proposition a.menu:focus,
-  #global-header #logo:focus {
-    color: #0b0c0c;
-  }
-  #global-header #logo:focus img {
-    filter: invert(1);
-  }
-  .account-menu__link:focus,
-  .is-smaller .account-menu__link--home:focus,
-  .is-smaller .account-menu__link--menu:focus,
-  .is-smaller .account-menu__link--home.account-menu__link--active:focus,
-  .is-smaller .account-menu__link--menu.account-menu__link--active:focus {
-    color: #0b0c0c;
-    background-color: #fd0;
-  }
-  @media (min-width: 641px){
-    #footer .footer-meta .copyright a {
-      background-position: 50% 0;
-      float: right;
-    }
-  }
-  #footer .footer-meta .copyright a:focus {
-    background-color: #fd0;
-  }
-
-  // govuk textarea
-  textarea:focus {
-    outline: 3px solid #fd0;
-    outline-offset: 0;
-    -webkit-box-shadow: inset 0 0 0 2px;
-    box-shadow: inset 0 0 0 2px;
-  }
-
-  // error focus state
-  .form-control.form-control-error:focus {
-    border-color: $input-border;
-    -webkit-box-shadow: none;
-    box-shadow: none
-  }
-
-  // radios
-  .multiple-choice input {
-    top: -2px;
-    left: -2px;
-    width: 44px;
-    height: 44px;
-  }
-  .multiple-choice [type="radio"] + label::before {
-    width: 40px;
-    height: 40px;
-    box-sizing: border-box;
-  }
-  .multiple-choice [type="radio"]:focus + label::before {
-    border-width: 4px;
-  }
-  .multiple-choice [type=radio]+label::after {
-    top: 10px;
-    left: 10px;
-  }
