--- app/assets/stylesheets/partials/_forms__mixin.scss
@@ -1,9 +1,0 @@
-@mixin form-control-sizes {
-  .form-control-1-4.form-control-1-4 {
-    width: 25%;
-  }
-
-  .form-control-1-3.form-control-1-3 {
-    width: 33%;
-  }
-}
