--- app/assets/stylesheets/partials/_modal.scss
@@ -1,12 +1,0 @@
-#timeout-sign-out-btn {
-  display: none;
-}
-
-#timeout-dialog h1 {
-  font-size: 24px;
-  line-height: 1.1em;
-
-  @media (min-width: 40.0625em){
-    font-size: 36px;
-  }
-}
