--- app/assets/stylesheets/partials/_page.scss
@@ -1,11 +1,0 @@
-.full-width .content__body {
-    float: none;
-    width: 100%;
-}
-
-@media screen and (min-width: 641px) {
-    .content__body {
-        float: left;
-        width: 61.666667%;
-    }
-}
