--- app/assets/stylesheets/partials/_shame.scss
@@ -1,16 +1,0 @@
-//See https://github.com/alphagov/govuk_elements/issues/221
-
-// Show the summary arrow in Firefox
-// The current Firefox implementation uses display: list-item to show the arrow marker.
-// We want to use display: inline-block to shrink-wrap the focus outline around the text.
-// This will turn off inline-block for Firefox only.
-// @-moz-document is going away: https://bugzilla.mozilla.org/show_bug.cgi?id=1035091 .
-// Hopefully they’ll fix <summary> first but if not it’ll fall back to no arrow:
-// https://bugzilla.mozilla.org/show_bug.cgi?id=1270163
-// If Mozilla add display: revert and remove list-item from summary then this will fall through.
-@-moz-document regexp('.*') {
-  details summary:not([tabindex]) {
-    display: list-item;
-    display: revert;
-  }
-}
