--- app/assets/stylesheets/partials/_task-list.scss
@@ -1,81 +1,0 @@
-// Task list pattern
-
-.app-task-list {
-  list-style-type: none;
-  padding-left: 0;
-  margin-top: 0;
-  margin-bottom: 0;
-  @include media(tablet) {
-    min-width: 550px;
-  }
-
-  .app-task-list__items {
-    padding-left: 0;
-  }
-}
-
-.app-task-list__section {
-  display: table;
-}
-
-.app-task-list__section-number {
-  display: table-cell;
-
-  @include media(tablet) {
-    min-width: govuk-spacing(6);
-    padding-right: 0;
-  }
-}
-
-.app-task-list__items {
-  font-family: "nta", Arial, sans-serif;
-  font-size: 19px;
-  @include govuk-responsive-margin(9, "bottom");
-  list-style: none;
-  padding-left: 0;
-  @include media(tablet) {
-    padding-left: govuk-spacing(6);
-  }
-}
-
-.app-task-list__item {
-  border-bottom: 1px solid $border-colour;
-  margin-bottom: 0 !important;
-  padding-top: govuk-spacing(2);
-  padding-bottom: govuk-spacing(2);
-  @include govuk-clearfix;
-}
-
-.app-task-list__item:first-child {
-  border-top: 1px solid $border-colour;
-}
-
-.app-task-list__task-name {
-  display: block;
-  @include media(450px) {
-    float: left;
-    width: 75%;
-  }
-}
-
-.app-task-list__task-completed {
-  margin-top: govuk-spacing(2);
-  margin-bottom: govuk-spacing(1);
-
-  @include media(450px) {
-    float: right;
-    margin-top: 0;
-    margin-bottom: 0;
-  }
-}
-
-.app-task-list__task-progress {
-  margin-top: govuk-spacing(2);
-  margin-bottom: govuk-spacing(1);
-
-  @include media(450px) {
-    float: right;
-    margin-top: 0;
-    margin-bottom: 0;
-  }
-}
