--- app/assets/stylesheets/partials/_template.scss
@@ -1,84 +1,0 @@
-/*
-====================================
-Fixes issues with missing styles when moving from AF 3.11 to 4.11
-====================================
- */
-
- html {
-    background: rgb(222, 224, 226);
-}
-body {
-    background: rgb(255, 255, 255);
-}
-.header__menu__proposition-name {
-    font-family: nta, Arial, sans-serif;
-    font-weight: 700;
-    text-transform: none;
-    font-size: 18px;
-    line-height: 1.2;
-    color: rgb(255, 255, 255);
-    text-decoration: none;
-}
-@media (min-width: 641px) {
-    .header__menu__proposition-name {
-        font-size: 24px;
-        line-height: 1.25;
-    }
-}
-
-.service-info {
-    display: block;
-    position: relative;
-    z-index: 997;
-    overflow: hidden;
-    padding-bottom: 0.5em;
-    border-top: 10px solid rgb(0, 94, 165);
-}
-
-.centered-content {
-    display: block;
-    max-width: 960px;
-    font-family: nta, Arial, sans-serif;
-    font-weight: 400;
-    text-transform: none;
-    font-size: 16px;
-    line-height: 1.25;
-    margin: 0px auto;
-    padding: 0em 0.78947em;
-}
-@media (min-width: 641px){
-    .centered-content {
-        padding: 0em 1.57895em;
-        font-size: 19px;
-        line-height: 1.31579;
-    }
-}
-
-
-.report-error {
-    display: inline-block;
-    clear: both;
-    width: 100%;
-    margin: 2em 0 1.5em 0;
-}
-@media (min-width: 641px){
-    .report-error {
-        width: 66%;
-        float: left;
-    }
-}
-.report-error .report-error__toggle {
-    font-family: "nta",Arial,sans-serif;
-    font-weight: 400;
-    text-transform: none;
-    font-size: 14px;
-    line-height: 1.14286;
-    color: #454a4c;
-}
-@media (min-width: 641px){
-    .report-error .report-error__toggle {
-        font-size: 16px;
-        line-height: 1.25;
-    }
-}
-
