--- app/assets/stylesheets/partials/_typography.scss
@@ -1,76 +1,0 @@
-/*
-====================================
-this fixes a specificity issue with content__body p selectors
-see https://github.com/hmrc/assets-frontend/issues/676
-====================================
-*/
-p.lede {
-  font-family: "nta",Arial,sans-serif;
-  font-size: 24px;
-  line-height: 1.25;
-  font-weight: 400;
-  text-transform: none;
-}
-/*replicate font-xsmall due to ".content p" rule*/
-p.font-xsmall{
-font-family:"nta",Arial,sans-serif;
-font-size:16px;
-line-height:1.25;
-font-weight:300;
-text-transform:none
-}
-@media (max-width: 640px){
-p.font-xsmall{
-  font-size:14px;
-  line-height:1.14286
-  }
-}
-
-
-/*give some space to the hero*/
-.hero {
-  margin: 1em 0;
-}
-.hero__heading-large,
-.hero__heading-small {
-  display: block;
-}
-
-
-/* CYA padding in line with current GDS */
-@media screen and (min-width: 769px) {
-  .govuk-check-your-answers > * > * {
-    padding-top: 0.53em;
-    padding-bottom: 0.53em;
-  }
-}
-
-.govuk-list, .app-prose-scope ul, .app-prose-scope ol {
-    font-family: "GDS Transport",Arial,sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    font-weight: 400;
-    line-height: 1.25;
-    color: #0b0c0c;
-    margin-top: 0;
-    margin-bottom: 15px;
-    padding-left: 0;
-    list-style-type: none;
-}
-
-.govuk-list--number, .app-prose-scope ol {
-    padding-left: 20px;
-    list-style-type: decimal;
-}
-
-.trust-list--bullet--padding-left {
-    padding-left: 20px;
-}
-
-.trust-list--bullet {
-  list-style-type: disc;
-}
-
-.trust-list--bullet .inside {
-  list-style-position: inside;
-}
