--- app/assets/stylesheets/partials/_utilities.scss
@@ -1,29 +1,0 @@
-/*
-====================================
-Utility alignments
-====================================
-*/
-.u-align--right {
-  text-align: right;
-}
-@media screen and (min-width: 641px){
-  .u-align--tablet-right {
-    text-align: right;
-  }
-}
-.u-align--center {
-    text-align: center;
-}
-
-/*
-====================================
-Utility bold
-Pulled from GDS elements typography
-====================================
-*/
-.bold {
-  font-weight: 700;
-}
-label > .bold {
-  display: block; /*temp fix until form story can fix label structure*/
-}
