--- app/assets/stylesheets/patterns/_add-to-a-list.scss
@@ -1,82 +1,0 @@
-.hmrc-add-to-a-list {
-  font-size: 19px;
-  table-layout: fixed;
-
-  @include media(tablet) {
-    display: table;
-  }
-}
-
-.hmrc-add-to-a-list--short {
-  @include media(tablet) {
-    // to make group of q&a line up horizontally (unless there is just one group)
-    width: 100%;
-    // recommended for mostly short questions
-    .hmrc-add-to-a-list__question {
-      width: 30%;
-    }
-  }
-}
-
-.hmrc-add-to-a-list--long {
-  @include media(tablet) {
-    // to make group of q&a line up horizontally (unless there is just one group)
-    width: 100%;
-    // recommended for mostly long questions
-    .hmrc-add-to-a-list__question {
-      width: 50%;
-    }
-  }
-}
-
-.hmrc-add-to-a-list__contents {
-  position: relative;
-  padding-bottom: 5px;
-  border-bottom: 1px solid $border-colour;
-
-  @include media(tablet) {
-    display: table-row;
-    padding-bottom: 0;
-    border-bottom-width: 0;
-  }
-}
-
-.hmrc-add-to-a-list__contents:first-child dt {
-  @include media(tablet) {
-    padding-top: 0;
-    padding-right: 20px;
-  }
-}
-
-.hmrc-add-to-a-list__identifier,
-.hmrc-add-to-a-list__remove,
-.hmrc-add-to-a-list__change {
-
-  position: relative;
-  display: block;
-  margin: 0;
-  text-align: left;
-  font-weight: normal;
-  overflow-wrap: break-word;
-  word-wrap: break-word;
-
-  @include media(tablet) {
-    display: table-cell;
-    padding: 15px 0;
-    border-bottom: 1px solid $border-colour;
-  }
-}
-
-.hmrc-add-to-a-list__identifier {
-  margin: 10px auto;
-
-  @include media(tablet) {
-    margin: 0;
-    padding-right: 20px;
-  }
-}
-
-.hmrc-add-to-a-list__identifier--light {
-  font-weight: normal;
-  padding-right: 20px;
-}
