--- app/assets/stylesheets/patterns/_check-your-answers.scss
@@ -1,8 +1,0 @@
-.cya-question, .cya-answer, .cya-change {
-    word-wrap: break-word;
-    overflow-wrap: break-word;
-}
-
-.govuk-check-your-answers {
-    table-layout: fixed;
-}
