--- app/assets/stylesheets/patterns/_symbol-input.scss
@@ -1,14 +1,0 @@
-.symbol {
-
-  .symbol-character {
-    width: 3%;
-    line-height: 33px;
-  }
-
-  .form-control, input[type="text"], input[type="password"] {
-    @include media(mobile) {
-      width: 97%;
-    }
-  }
-}
-
