+++ app/views/AddAProtectorView.scala.html
@@ -20,7 +20,7 @@
 @import viewmodels.RadioOption
 
 @this(
-    main_template: MainTemplate,
+    mainTemplate: MainTemplate,
     formHelper: FormWithCSRF
 )
 
@@ -32,7 +32,7 @@
     maxedOut: List[String]
 )(implicit request: Request[_], messages: Messages)
 
-@main_template(
+@mainTemplate(
     title = s"${errorPrefix(form)}$heading"
 ) {
