--- app/views/GovukWrapper.scala.html
@@ -1,144 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import play.twirl.api.HtmlFormat
-@import views.html.layouts.GovUkTemplate
-
-@this(
-    header_nav: HeaderNav,
-    head: HeadWithTrackingConsent,
-    footer: Footer,
-    serviceInfo: ServiceInfo,
-    main_content_header: MainContentHeader,
-    footer_links: FooterLinks,
-    main_content: MainContent,
-    reportAProblemLink: ReportAProblemLink,
-    hmrcGovUkTemplate: GovUkTemplate
-)
-
-@(
-    appConfig: config.FrontendAppConfig,
-    title: String,
-    mainClass: Option[String] = None,
-    mainDataAttributes: Option[Html] = None,
-    bodyClasses: Option[String] = None,
-    sidebar: Html = HtmlFormat.empty,
-    contentHeader: Option[Html] = None,
-    mainContent: Html = HtmlFormat.empty,
-    serviceInfoContent: Html = HtmlFormat.empty,
-    scriptElem: Option[Html] = None,
-    canSignOut: Boolean = true
-    )(implicit request: Request[_], messages: Messages)
-
-@headContent = {
-    @head(
-        linkElem = None,
-        headScripts = None
-    )
-    <meta name="format-detection" content="telephone=no" />
-    <link rel="stylesheet" href='@routes.Assets.versioned("stylesheets/maintainprotectorsfrontend-app.css")' />
-}
-
-@headerNavLinks = {
-    @if(canSignOut) {
-        <li>
-            <a id="logOut" href="@controllers.routes.LogoutController.logout" data-target="link - click:@title:@Messages("site.sign_out")">
-            @Messages("site.sign_out")
-            </a>
-        </li>
-    }
-}
-
-@insideHeader = {
-    @header_nav(
-        navTitle = Some(messages("site.service_name")),
-        navTitleLink = None,
-        showBetaLink = false,
-        navLinks = Some(headerNavLinks)
-    )
-}
-
-@afterHeader = {}
-
-@bodyEnd = {
-    @footer(
-        analyticsToken = Some(appConfig.analyticsToken),
-        analyticsHost = "",
-        ssoUrl = None,
-        scriptElem = scriptElem,
-        gaCalls = None
-    )
-    <script src='@routes.Assets.versioned("javascripts/maintainprotectorsfrontend-app.min.js")'> </script>
-}
-
-@footerTop = {}
-
-@footerLinks = {
-    @footer_links(
-        additionalLinks = None,
-        euExitLinks = None,
-        accessibilityFooterUrl = None
-    )
-}
-
-@banner = {
-    @betaBanner(
-        userLoggedIn = true,
-        betaFeedbackUrl = appConfig.betaFeedbackUrl,
-        betaFeedbackUnauthenticatedUrl = appConfig.betaFeedbackUnauthenticatedUrl,
-        showFeedbackLinkInBetaBanner = true
-    )
-}
-
-@serviceInfoBlock = {
-    @serviceInfo(
-        betaBanner = banner,
-        includeHMRCBranding = false,
-        includeGridWrapper = false,
-        serviceInfoContent = Some(serviceInfoContent)
-    )
-}
-
-@mainContentHeader = {
-
-    @if(appConfig.languageTranslationEnabled) {
-        @views.html.language_selection(
-            appConfig.languageMap,
-            appConfig.routeToSwitchLanguage,
-            Some("translate u-align--right")
-        )
-    }
-
-    @if(contentHeader.isDefined) {
-        @main_content_header(contentHeader = contentHeader.get)
-    }
-}
-
-@getHelpForm = @{reportAProblemLink(appConfig.reportAProblemPartialUrl, appConfig.reportAProblemNonJSUrl)}
-
-@content = {
-    @main_content(
-        article = mainContent,
-        mainClass = mainClass,
-        mainDataAttributes = mainDataAttributes,
-        mainContentHeader = mainContentHeader,
-        serviceInfo = serviceInfoBlock,
-        getHelpForm = getHelpForm,
-        sidebar = sidebar
-    )
-}
-
-@hmrcGovUkTemplate(Some(title), bodyClasses)(headContent, bodyEnd, insideHeader, afterHeader, footerTop, Some(footerLinks), true)(content)
