+++ app/views/resources/Stylesheets.scala.html
@@ -14,21 +14,9 @@
  * limitations under the License.
  *@
 
-@import config.FrontendAppConfig
+@this()
 
-@this(
-    mainTemplate: MainTemplate,
-    appConfig: FrontendAppConfig
-)
-
-@(pageTitle: String, heading: String, message: String)(implicit request: Request[_], messages: Messages)
-
-@contentHeader = {
-  @components.heading(heading)
-}
-
-@mainContent = {
-  <p>@message</p>
-}
-
-@mainTemplate(title = pageTitle, contentHeader = Some(contentHeader))(mainContent = mainContent)
+@()
+<link href='@controllers.routes.Assets.versioned("stylesheets/application.css")' media="all" rel="stylesheet" type="text/css" />
+<script src="https://code.jquery.com/jquery-3.6.0.min.js" integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
+<script src='@routes.Assets.versioned("javascripts/maintainprotectorsfrontend-app.min.js")'> </script>
