+++ build.sbt
@@ -23,8 +23,10 @@ lazy val root = (project in file("."))
     TwirlKeys.templateImports ++= Seq(
       "play.twirl.api.HtmlFormat",
       "play.twirl.api.HtmlFormat._",
-      "uk.gov.hmrc.play.views.html.helpers._",
-      "uk.gov.hmrc.play.views.html.layouts._",
+      "uk.gov.hmrc.govukfrontend.views.html.components._",
+      "uk.gov.hmrc.govukfrontend.views.html.helpers._",
+      "uk.gov.hmrc.hmrcfrontend.views.html.components._",
+      "uk.gov.hmrc.hmrcfrontend.views.html.helpers._",
       "views.ViewUtils._",
       "models.Mode",
       "controllers.routes._"
@@ -49,10 +51,11 @@ lazy val root = (project in file("."))
     Concat.groups := Seq(
       "javascripts/maintainprotectorsfrontend-app.js" ->
         group(Seq(
-          "javascripts/show-hide-content.js",
           "javascripts/maintainprotectorsfrontend.js",
           "javascripts/autocomplete.js",
-          "javascripts/libraries/location-autocomplete.min.js"
+          "javascripts/libraries/location-autocomplete.min.js",
+          "javascripts/iebacklink.js",
+          "javascripts/print.js"
         ))
     ),
     // prevent removal of unused code which generates warning errors due to use of third-party libs
