+++ conf/prod.routes
@@ -1,6 +1,5 @@
 # Add all the application routes to the app.routes file
 ->         /maintain-a-trust/protectors                    app.Routes
 ->         /                          health.Routes
-->     	   /template                  template.Routes
 
 GET        /admin/metrics             @com.kenshoo.play.metrics.MetricsController.metrics
